@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace MondayFactory\DatabaseModelGenerator\Command;
 
+use Exception;
+use InvalidArgumentException;
 use MondayFactory\DatabaseModelGenerator\Generator\CollectionGenerator;
 use MondayFactory\DatabaseModelGenerator\Generator\DataGenerator;
 use MondayFactory\DatabaseModelGenerator\Generator\LowLevelDatabaseStorageGenerator;
@@ -16,63 +18,43 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Throwable;
+use UnexpectedValueException;
 
 class GenerateBasicModel extends Command
 {
 
 	public const COMMAND_NAME = 'basic-model';
 
+	/**
+	 * @var string
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.PropertyTypeHint.MissingAnyTypeHint
+	 * @phpcsSuppress SlevomatCodingStandard.TypeHints.PropertyTypeHint.MissingNativeTypeHint
+	 */
 	protected static $defaultName = 'basic-model';
 
-	/**
-	 * @var array
-	 */
-	private $definition;
+	/** @var array<scalar, mixed> */
+	private array $definition;
 
-	/**
-	 * @var string
-	 */
-	private $neonName;
+	private string $neonName;
 
-	/**
-	 * @var array
-	 */
-	private $generators = ['collection', 'data', 'llstorage'];
+	/** @var array<int, string> */
+	private array $generators = ['collection', 'data', 'llstorage'];
 
-	/**
-	 * @var bool
-	 */
-	private $printToStdOut;
+	private bool $printToStdOut;
 
-	/**
-	 * @var bool
-	 */
-	private $dryRun;
+	private bool $dryRun;
 
-	/**
-	 * @var bool
-	 */
-	private $force;
+	private bool $force;
 
-	/**
-	 * @var OutputInterface
-	 */
-	private $output;
+	private OutputInterface $output;
 
-	/**
-	 * @var string
-	 */
-	private $projectFilesPath;
+	private string $projectFilesPath;
 
-	/**
-	 * @var InputInterface
-	 */
-	private $input;
+	private InputInterface $input;
 
-	/**
-	 * @var array
-	 */
-	private $errors = [];
+	/** @var array<int, string> */
+	private array $errors = [];
 
 	public function __construct()
 	{
@@ -92,10 +74,10 @@ class GenerateBasicModel extends Command
 			getcwd() . '/app/model');
 		$this->addArgument(
 			'whatGenerate',
-			InputArgument::IS_ARRAY | InputArgument::OPTIONAL ,
+			InputArgument::IS_ARRAY | InputArgument::OPTIONAL,
 			'What You need to generate? You can add more types as list separated by space. 
 				For all leave argument blank. Allowed values: 
-				[' . join(', ', $this->generators) . ']' , ['collection', 'data', 'llstorage']);
+				[' . join(', ', $this->generators) . ']', ['collection', 'data', 'llstorage']);
 		$this->addOption('def-dir', null, InputOption::VALUE_REQUIRED, '', getcwd() . '/modelDefinition');
 		$this->addOption('ignored-namespace', 'I', InputOption::VALUE_REQUIRED, '', null);
 		$this->addOption('p', 'p', InputOption::VALUE_NONE, 'Print generated files to stdout');
@@ -103,28 +85,19 @@ class GenerateBasicModel extends Command
 		$this->addOption('force', 'f', InputOption::VALUE_NONE);
 	}
 
-	/**
-	 * @throws \InvalidArgumentException
-	 */
 	protected function execute(InputInterface $input, OutputInterface $output): int
 	{
 		$this->input = $input;
 
 		if (is_string($input->getArgument('projectFilesPath'))) {
-			$this->projectFilesPath = $input->getArgument('projectFilesPath');
+			$this->projectFilesPath = $input->getArgument('projectFilesPath') . '';
 		} else {
-			throw new \InvalidArgumentException('Parameter projectFilesPath must be string.');
+			throw new InvalidArgumentException('Parameter projectFilesPath must be string.');
 		}
 
-		$this->printToStdOut = is_bool($input->getOption('p'))
-			? $input->getOption('p')
-			: false;
-		$this->dryRun = is_bool($input->getOption('dry-run'))
-			? $input->getOption('dry-run')
-			: false;
-		$this->force = is_bool($input->getOption('force'))
-			? $input->getOption('force')
-			: false;
+		$this->printToStdOut = (bool) $input->getOption('p');
+		$this->dryRun = (bool) $input->getOption('dry-run');
+		$this->force = (bool) $input->getOption('force');
 		$this->output = $output;
 
 		if ($this->dryRun) {
@@ -132,21 +105,20 @@ class GenerateBasicModel extends Command
 		}
 
 		try {
-			if (! is_string($input->getArgument('neonName'))) {
-				throw new \InvalidArgumentException('Neon name must be string.');
+			if (!is_string($input->getArgument('neonName'))) {
+				throw new InvalidArgumentException('Neon name must be string.');
 			}
 
-			$this->neonName = $input->getArgument('neonName');
-			$neon = $this->resolveNenonPath();
+			$this->neonName = $input->getArgument('neonName') . '';
+			$neon = $this->resolveNenoPath();
 			$neonContent = file_get_contents($neon);
 
 			if ($neonContent === false) {
-				throw new \UnexpectedValueException('Neon file can not be read.');
+				throw new UnexpectedValueException('Neon file can not be read.');
 			}
 
 			$this->definition = Neon::decode($neonContent);
-
-		} catch (\Exception $e) {
+		} catch (Throwable $e) {
 			$output->writeln($e->getMessage());
 
 			return 1;
@@ -157,7 +129,7 @@ class GenerateBasicModel extends Command
 			: $this->generators;
 
 		foreach ($whatGenerate as $generator) {
-			if (! in_array($generator, $this->generators, true)) {
+			if (!in_array($generator, $this->generators, true)) {
 				$output->writeln("Generator {$generator} is not allowed. Skipping.");
 
 				continue;
@@ -168,10 +140,12 @@ class GenerateBasicModel extends Command
 					$this->generateCollection();
 
 					break;
+
 				case 'data':
 					$this->generateData();
 
 					break;
+
 				case 'llstorage':
 					$this->generateLlstorage();
 
@@ -179,13 +153,12 @@ class GenerateBasicModel extends Command
 			}
 		}
 
-		if (count($this->errors)) {
+		if ($this->errors !== []) {
 			$this->output->writeln("<fg=yellow>Finished with errors.</> ");
 
 			foreach ($this->errors as $error) {
 				$this->output->writeln("<fg=yellow>$error</> ");
 			}
-
 		} else {
 			$this->output->writeln("<fg=green>Done.</>");
 		}
@@ -196,7 +169,7 @@ class GenerateBasicModel extends Command
 	private function generateCollection(): void
 	{
 		$collectionGenerator = new CollectionGenerator($this->definition, $this->getNeonName());
-		$printer = new Printer();
+		$printer = new Printer;
 
 		$this->processOutput($printer->printFile($collectionGenerator->file), $collectionGenerator->getFileNamespace());
 	}
@@ -204,7 +177,7 @@ class GenerateBasicModel extends Command
 	private function generateData(): void
 	{
 		$dataGenerator = new DataGenerator($this->definition, $this->getNeonName());
-		$printer = new Printer();
+		$printer = new Printer;
 
 		$this->processOutput($printer->printFile($dataGenerator->file), $dataGenerator->getFileNamespace());
 	}
@@ -212,9 +185,12 @@ class GenerateBasicModel extends Command
 	private function generateLlstorage(): void
 	{
 		$lowLevelDatabaseStorage = new LowLevelDatabaseStorageGenerator($this->definition, $this->getNeonName());
-		$printer = new Printer();
+		$printer = new Printer;
 
-		$this->processOutput($printer->printFile($lowLevelDatabaseStorage->file), $lowLevelDatabaseStorage->getFileNamespace());
+		$this->processOutput(
+			$printer->printFile($lowLevelDatabaseStorage->file),
+			$lowLevelDatabaseStorage->getFileNamespace(),
+		);
 	}
 
 	private function processOutput(string $output, string $filePath): void
@@ -226,40 +202,32 @@ class GenerateBasicModel extends Command
 	{
 		$namespace = ltrim(str_replace('\\', '/', $namespace), '/');
 
-		if (is_string($this->input->getOption('ignored-namespace')))
-		{
+		if (is_string($this->input->getOption('ignored-namespace'))) {
 			$ignoredFileSpace = str_replace('\\', '/', $this->input->getOption('ignored-namespace'));
 			$ignoredFileSpace = rtrim(ltrim($ignoredFileSpace, '/'), '/');
 
 			$namespace = ltrim(substr($namespace, strlen($ignoredFileSpace)), '/');
 		}
 
-		$projectFilesPath = '';
-
-		if (strlen($this->projectFilesPath) > 0) {
-			$projectFilesPath = substr($this->projectFilesPath, 0, 1) === '/'
-				? realpath($this->projectFilesPath)
-				: realpath(getcwd() . '/' . $this->projectFilesPath);
-
-			if (strlen($this->projectFilesPath) > 1 && $projectFilesPath === false) {
-			}
+		if ($this->projectFilesPath !== '') {
+			$projectFilesPath = $this->projectFilesPath[0] === '/' ? realpath($this->projectFilesPath) : realpath(
+                getcwd() . '/' . $this->projectFilesPath,
+            );
 		} else {
 			$cwd = getcwd();
-			$projectFilesPath = $cwd !== false
-				? realpath($cwd)
-				: false;
+			$projectFilesPath = $cwd !== false ? realpath($cwd) : false;
 		}
 
-		$projectFilesPath = $projectFilesPath === false
-			? ''
-			: $projectFilesPath;
+		$projectFilesPath = $projectFilesPath === false ? '' : $projectFilesPath;
 
 		$fileName = $projectFilesPath . '/' . $namespace . '.php';
 
-
 		if ($this->dryRun) {
 			if (file_exists($fileName)) {
-				$this->output->write("<fg=red>File {$fileName} exists. You must use force (--force|-f) to recreate.</>", true);
+				$this->output->write(
+					"<fg=red>File {$fileName} exists. You must use force (--force|-f) to recreate.</>",
+					true,
+				);
 			} elseif ($projectFilesPath === '') {
 				$this->errors[] = "Cannot save file $fileName, directory $this->projectFilesPath is not exists!";
 
@@ -268,30 +236,30 @@ class GenerateBasicModel extends Command
 				$this->output->write("<fg=green>File {$fileName} is ready for create.</>", true);
 			}
 		}
+
 		if ($this->printToStdOut) {
-			if (! $this->dryRun) {
+			if (!$this->dryRun) {
 				$this->output->write("<fg=green>{$fileName}</>", true);
 			}
 
 			try {
 				$oldFileContent = @file_get_contents($fileName);
 
-				if (false === $oldFileContent) {
+				if ($oldFileContent === false) {
 					$this->output->writeln($content);
 				} else {
 					require_once 'SideBySide.php';
 
-					$diff = new \SideBySide();
-					$diff->setAffixes(\SideBySide::AFFIXES_CLI);
+					$diff = new SideBySide;
+					$diff->setAffixes(SideBySide::AFFIXES_CLI);
 					echo $diff->compute($oldFileContent, $content)[1];
 				}
-
-			} catch (\Exception $e) {
+			} catch (Throwable $e) {
 				$this->output->writeln($content);
 			}
-
 		}
-		if (! $this->dryRun) {
+
+		if (!$this->dryRun) {
 			if ($projectFilesPath === '') {
 				$this->errors[] = "Cannot save file $fileName, directory $this->projectFilesPath is not exists!";
 
@@ -305,20 +273,26 @@ class GenerateBasicModel extends Command
 					$this->output->write("<fg=red>File {$fileName} cannot be written.</> " . $e->getMessage(), true);
 				}
 			} else {
-				if (! file_exists($fileName)) {
+				if (!file_exists($fileName)) {
 					try {
 						FileSystem::write($fileName, $content);
 					} catch (IOException $e) {
-						$this->output->write("<fg=red>File {$fileName} cannot be written.</> " . $e->getMessage(), true);
+						$this->output->write(
+							"<fg=red>File {$fileName} cannot be written.</> " . $e->getMessage(),
+							true,
+						);
 					}
 				} else {
-					$this->output->write("<fg=red>File {$fileName} exists. Use force (--force|-f) if you need recreate the model.</>", true);
+					$this->output->write(
+						"<fg=red>File {$fileName} exists. Use force (--force|-f) if you need recreate the model.</>",
+						true,
+					);
 				}
 			}
 		}
 	}
 
-	private function resolveNenonPath(): string
+	private function resolveNenoPath(): string
 	{
 		$path = __DIR__ . '/../../../../../modelDefinition/';
 
@@ -328,12 +302,11 @@ class GenerateBasicModel extends Command
 
 		$filePath = realpath($path . '/' . $this->getNeonName() . '.neon');
 
-		if ($filePath !== false && file_exists($filePath))
-		{
+		if ($filePath !== false && file_exists($filePath)) {
 			return $filePath;
 		}
 
-		throw new \InvalidArgumentException("File {$path}/{$this->getNeonName()}.neon not found.");
+		throw new InvalidArgumentException("File {$path}/{$this->getNeonName()}.neon not found.");
 	}
 
 	private function getNeonName(): string
@@ -341,7 +314,7 @@ class GenerateBasicModel extends Command
 		$result = preg_replace('/.neon/', '', $this->neonName);
 
 		if (is_null($result)) {
-			throw new \Exception('Some error occured.');
+			throw new Exception('Some error occured.');
 		}
 
 		return $result;
