@@ -4,21 +4,20 @@ declare(strict_types=1);
 
 namespace MondayFactory\DatabaseModelGenerator\Generator;
 
-use MondayFactory\DatabaseModel\Colection\BaseDatabaseDataCollection;
-use MondayFactory\DatabaseModel\Colection\IDatabaseDataCollection;
+use MondayFactory\DatabaseModel\Collection\BaseDatabaseDataCollection;
+use Nette\PhpGenerator\ClassType;
 use Nette\PhpGenerator\PhpFile;
 use Nette\PhpGenerator\PhpLiteral;
+use ReflectionClass;
 
 class CollectionGenerator
 {
 
 	use TBaseMethods;
 
-	/**
-	 * @var PhpFile
-	 */
-	public $file;
+	public PhpFile $file;
 
+	/** @param array<scalar, mixed> $definition */
 	public function __construct(array $definition, string $name)
 	{
 		$this->definition = $definition;
@@ -29,13 +28,12 @@ class CollectionGenerator
 
 	private function generate(): PhpFile
 	{
-		$file = new PhpFile();
+		$file = new PhpFile;
 		$file->setStrictTypes();
 
 		$namespace = $file->addNamespace($this->getNamespace('Collection'));
 		$namespace
 			->addUse(BaseDatabaseDataCollection::class)
-			->addUse(IDatabaseDataCollection::class)
 			->addUse($this->getRowFactoryNamespace() . '\\' . $this->getRowFactoryClassName());
 
 		$class = $namespace->addClass($this->getCollectionFactoryClassName());
@@ -43,57 +41,93 @@ class CollectionGenerator
 
 		$class->setExtends(BaseDatabaseDataCollection::class);
 
-		if (isset($this->definition['databaseTableId']) && ! empty($this->definition['databaseTableId'])) {
-			$idFieldSerializerProperty = $class->addProperty('idFieldSerializer');
+		if (isset($this->definition['databaseTableId']) && (string) $this->definition['databaseTableId'] !== '') {
+			$idFieldSerializerProperty = $class->addProperty('idFieldSerializer')
+				->setType('string')
+				->setNullable();
 
 			$idField = $this->findIdField();
 
 			if (is_array($idField) && isset($idField['toString'])) {
-
-				$idFieldSerializer = preg_replace("/(\(.*\))/m",'', $idField['toString']);
-				$idFieldSerializer = preg_replace("/(->)+/m",'', $idFieldSerializer);
+				$idFieldSerializer = preg_replace("/(\(.*\))/m", '', $idField['toString']);
+				$idFieldSerializer = preg_replace("/(->)+/m", '', $idFieldSerializer);
 
 				$idFieldSerializerProperty->setValue($idFieldSerializer)->setProtected();
 			}
 		}
 
-		$methodCreate = $class->addMethod('create');
+		$this->addCreateMethod($class);
 
-		$methodCreate->addBody('return new static($data, ?, $idField);', [new PhpLiteral($this->getRowFactoryClassName() . '::class')])
-			->setVisibility('public')
-			->setStatic()
-			->setReturnType(IDatabaseDataCollection::class)
-			->addComment("@param array \$data\n")
-			->addComment('@return ' . (new \ReflectionClass(IDatabaseDataCollection::class))->getShortName())
-			->addParameter('data')
-			->setTypeHint('iterable');
+		$rowFactoryClassReflection = new ReflectionClass(
+			$this->getRowFactoryNamespace() . '\\' . $this->getRowFactoryClassName(),
+		);
 
-		$methodCreate->addParameter('idField')
-			->setTypeHint('string')
-			->setNullable()
-			->setDefaultValue(null);
+		$this->addGetByKeyMethod($class, $rowFactoryClassReflection);
+		$this->addCurrentMethod($class, $rowFactoryClassReflection);
 
 		return $file;
 	}
 
+	/** @return array<scalar, mixed>|null */
 	private function findIdField(): ?array
 	{
-		if (
-			isset($this->definition['databaseCols']['rw'])
-			&& isset(
-				$this->definition['databaseCols']['rw'][$this->definition['databaseTableId']]
-			)
-		) {
-			return $this->definition['databaseCols']['rw'][$this->definition['databaseTableId']];
-		} else if (isset($this->definition['databaseCols']['rw'])
-			&& isset(
-				$this->definition['databaseCols']['ro'][$this->definition['databaseTableId']]
-			)
-		) {
-			return $this->definition['databaseCols']['ro'][$this->definition['databaseTableId']];
-		}
+		return $this->definition['databaseCols']['rw'][$this->definition['databaseTableId']]
+			   ?? $this->definition['databaseCols']['ro'][$this->definition['databaseTableId']]
+				  ?? null;
+	}
 
-		return null;
+	private function addCreateMethod(ClassType $class): void
+	{
+		$method = $class->addMethod('create');
+		$method->addComment('@param iterable<int|string, mixed> $data');
+
+		$method->addBody(
+			'return new self($data, ?, $idField);',
+			[new PhpLiteral($this->getRowFactoryClassName() . '::class')],
+		)
+			->setVisibility('public')
+			->setStatic()
+			->setReturnType('self')
+			->addParameter('data')
+			->setType('iterable');
+
+		$method->addParameter('idField')
+			->setType('string')
+			->setNullable()
+			->setDefaultValue(null);
+	}
+
+    private function addGetByKeyMethod(ClassType $class, ReflectionClass $rowFactoryClassReflection): void
+	{
+		$method = $class->addMethod('getByKey');
+		$method->addComment('@param int|string $key');
+
+		$method->addBody('$data = parent::getByKey($key);')
+			->addBody(
+				'assert($data === null || $data instanceOf ?);',
+				[new PhpLiteral($this->getRowFactoryClassName())],
+			)
+			->addBody('')
+			->addBody('return $data;')
+			->setVisibility('public')
+			->setReturnNullable()
+			->setReturnType($rowFactoryClassReflection->getName());
+
+		$method->addParameter('key');
+	}
+
+    private function addCurrentMethod(ClassType $class, ReflectionClass $rowFactoryClassReflection): void
+	{
+		$method = $class->addMethod('current');
+		$method->addBody('$data = parent::current();')
+			->addBody(
+				'assert($data === false || $data instanceOf ?);',
+				[new PhpLiteral($this->getRowFactoryClassName())],
+			)
+			->addBody('')
+			->addBody('return $data;')
+			->setComment('@return false|' . $rowFactoryClassReflection->getShortName())
+			->setVisibility('public');
 	}
 
 }
